% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom}
\alias{binom}
\title{Binomial distribution as table.}
\usage{
binom(times, prob_success)
}
\arguments{
\item{times}{number of trials}

\item{prob_success}{probability of success (number between 0 and 1)}
}
\value{
Binomial distribution as a tibble
}
\description{
Generates a tibble containing the binomial distribution using dbinom().
}
\examples{
binom(times = 10, prob_success = 1/10)
}
