% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dice.R
\name{circle_points}
\alias{circle_points}
\title{Helper function to draw a circle}
\usage{
circle_points(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{Vector with x and y coordinate of center}

\item{diameter}{Diameter of circle}

\item{npoints}{Number of points used for drawing a circle}
}
\value{
Dataframe with x and y coordinates to draw a circle
}
\description{
Helper function to draw a circle
}
\examples{
circle_points()
}
