% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-eval.R
\name{eval_dice_expression}
\alias{eval_dice_expression}
\title{Evaluate a parsed dice expression}
\usage{
eval_dice_expression(node, n, prob = NULL, detail = FALSE)
}
\arguments{
\item{node}{A tree returned by parse_dice_expression()}

\item{n}{Number of times the expression is evaluated}

\item{prob}{Probabilities of the sides of the dice, see resolve_prob()}

\item{detail}{If TRUE, keep the dice that were rolled (in the attribute \code{dice})}
}
\value{
Numeric vector of length n. With \code{detail}, it has an attribute \code{dice}:
a list of n tibbles (group, sides, value, kept), one row for each die.
Dice that are not \code{kept} were dropped by a modifier.
}
\description{
Evaluate a parsed dice expression
}
