% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydice.R
\name{force_coin}
\alias{force_coin}
\title{Force a coin flipping result.}
\usage{
force_coin(data = NULL, result = 6, round = 1, experiment = 1,
  success = 2)
}
\arguments{
\item{data}{Data from a previous experiment}

\item{result}{Vector of flipping coin results}

\item{round}{Round of flipping coin}

\item{experiment}{Experiment Number}

\item{success}{Which result is a success (default = 6)}
}
\value{
Result of experiment as a tibble
}
\description{
The forced result is returned as a tibble.
}
\examples{
force_coin(6)
force_coin(1:6)
}
