% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydice.R
\name{force_dice}
\alias{force_dice}
\title{Force a dice rolling result.}
\usage{
force_dice(data = NULL, result = 6, round = 1, experiment = 1,
  success = 6)
}
\arguments{
\item{data}{Data from a previous experiment}

\item{result}{Vector of rolling dice results}

\item{round}{Round of rolling dice}

\item{experiment}{Experiment Number}

\item{success}{Which result is a success (default = 6)}
}
\value{
Result of experiment as a tibble
}
\description{
The forced result is returned as a tibble.
}
\examples{
force_dice(6)
force_dice(1:6)
}
