% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-eval.R
\name{parse_dice_expression}
\alias{parse_dice_expression}
\title{Parse a dice expression}
\usage{
parse_dice_expression(dice_formula)
}
\arguments{
\item{dice_formula}{A string containing a dice formula, e.g. 4d6e6kh3+2}
}
\value{
A tree (nested lists) describing the expression
}
\description{
Parse a dice expression
}
