% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-formula.R
\name{parse_dice_formula_part}
\alias{parse_dice_formula_part}
\title{Helper function to parse a dice formula}
\usage{
parse_dice_formula_part(dice_formula_part)
}
\arguments{
\item{dice_formula_part}{A split dice formula, e.g. 1d6e2. For more complex
formula, e.g. 1d6e2+3d4, see parse_dice_formula}
}
\description{
Helper function to parse a dice formula
}
