% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-formula.R
\name{parse_explode}
\alias{parse_explode}
\title{Helper function to parse the exploding dice modifier of a dice formula}
\usage{
parse_explode(dice_formula, dice_sides)
}
\arguments{
\item{dice_formula}{A string containing a dice formula, e.g. 1d6e>5}

\item{dice_sides}{Number of sides of the dice}
}
\value{
Integer vector of the sides that make a die explode (empty if the
formula has no exploding modifier)
}
\description{
Supports the syntax of Avrae's d20 library: \code{e} (explode on the highest
side), \code{e6} (explode on a 6), \code{e>5} (explode on 5 or more) and \code{e<2}
(explode on 2 or less).
}
