% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{plot_binom}
\alias{plot_binom}
\title{Plot a binomial distribution.}
\usage{
plot_binom(data, title = "Binomial distribution", color = "darkgrey",
  color_highlight = "red", label = NULL, label_size = 3,
  min_pct = 0.05, highlight = NULL)
}
\arguments{
\item{data}{data containing values for binomial distribution}

\item{title}{title of the plot}

\item{color}{color of bars}

\item{color_highlight}{color of highlighted bars}

\item{label}{add labels to plot?}

\item{label_size}{size of label}

\item{min_pct}{surpress values < min_pct}

\item{highlight}{vector of values to be highlighted}
}
\value{
ggplot object
}
\description{
Plot a binomial distribution generated with dice_binom() or coin_binom()
}
\examples{
plot_binom(data = binom_dice(times = 10))
}
