% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin-plot.R
\name{plot_coin}
\alias{plot_coin}
\title{Plot result of flip_coin()}
\usage{
plot_coin(
  data,
  detailed = FALSE,
  fill = "white",
  fill_success = "gold",
  line_color = "black",
  line_size = 0.8,
  show_result = TRUE
)
}
\arguments{
\item{data}{result of flip_coin()}

\item{detailed}{not supported at moment}

\item{fill}{Fill color}

\item{fill_success}{Fill color if result is a success}

\item{line_color}{Color of Lines}

\item{line_size}{Size of Lines}

\item{show_result}{If TRUE (default), the result is written on each coin}
}
\value{
ggplot-Object
}
\description{
Plot result of flip_coin()
}
\examples{
library(magrittr)  

# plot one coin
plot_coin()

# plot multiple coin flips
flip_coin(times = 3, rounds = 3) \%>\% 
  plot_coin()
  
# change coin design    
flip_coin(times = 3, rounds = 3) \%>\% 
  plot_coin(fill_success = "red")
}
