% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin-plot.R
\name{plot_single_coin}
\alias{plot_single_coin}
\title{Draw a single coin}
\usage{
plot_single_coin(
  ggplot = NULL,
  result = 1,
  x = 0,
  y = 0,
  width = 0.9,
  fill = "white",
  detailed = FALSE,
  line_size = 0.8,
  line_color = "black",
  show_result = FALSE
)
}
\arguments{
\item{ggplot}{ggplot-Object. If passed, the dice will be added to plot}

\item{result}{Result of flip coin}

\item{x}{X-coordinate of dice (center)}

\item{y}{y-coordinate of dice (center)}

\item{width}{Width of coin}

\item{fill}{Fill color}

\item{detailed}{If TRUE, the dice is plotted with more details}

\item{line_size}{Size of Lines}

\item{line_color}{Color of Lines}

\item{show_result}{If TRUE, the result is written on the coin}
}
\value{
ggplot-Object
}
\description{
Draw a single coin
}
