% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-eval.R
\name{resolve_prob}
\alias{resolve_prob}
\title{Probabilities of the sides, for each group of dice of an expression}
\usage{
resolve_prob(node, prob = NULL)
}
\arguments{
\item{node}{A tree returned by parse_dice_expression()}

\item{prob}{NULL (fair dice), a vector of probabilities that is used for
every group of dice, or a list with a vector (or NULL) for each group of
dice, from left to right}
}
\value{
A list with one element for each group of dice
}
\description{
Probabilities of the sides, for each group of dice of an expression
}
