% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-formula.R
\name{top_n_dice}
\alias{top_n_dice}
\title{Helper function to get sum of top n dice}
\usage{
top_n_dice(x, n, dec = F)
}
\arguments{
\item{x}{Vector of dice-values}

\item{n}{Number of dice}
}
\description{
Helper function to get sum of top n dice
}
