% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_coin}
\alias{binom_coin}
\title{Binomial distribution of flipping a coin.}
\usage{
binom_coin(times, sides = 2, success = 2)
}
\arguments{
\item{times}{how many times a coin is flipped (or how many coins are flipped at the same time)}

\item{sides}{number of sides of the coin (default = 2)}

\item{success}{which result is a success (default = 2)}
}
\value{
binomial distribution as a tibble
}
\description{
Generates a tibble containing the binomial distribution of flipping a coin
}
\examples{
binom_coin(times = 10)
}
