% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_dice}
\alias{binom_dice}
\title{Binomial distribution of rolling a dice.}
\usage{
binom_dice(times, sides = 6, success = 6)
}
\arguments{
\item{times}{How many times a dice is rolled (or how many dice are rolled at the same time)}

\item{sides}{Number of sides of the dice (default = 6)}

\item{success}{Which result is a success (default = 6)}
}
\value{
Binomial distribution as a tibble
}
\description{
Generates a tibble containing the binomial distribution of rolling the dice.
}
\examples{
binom_dice(times = 10)
}
