% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-eval.R
\name{eval_dice_expression}
\alias{eval_dice_expression}
\title{Evaluate a parsed dice expression}
\usage{
eval_dice_expression(node, n, prob = NULL)
}
\arguments{
\item{node}{A tree returned by parse_dice_expression()}

\item{n}{Number of times the expression is evaluated}

\item{prob}{Probabilities of the sides of the dice, see resolve_prob()}
}
\value{
Numeric vector of length n
}
\description{
Evaluate a parsed dice expression
}
