% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dice.R
\name{plot_dice}
\alias{plot_dice}
\title{Plot result of roll_dice()}
\usage{
plot_dice(data, detailed = TRUE, fill = "white",
  fill_success = "green", point_color = "black",
  line_color = "black")
}
\arguments{
\item{data}{result of roll_dice()}

\item{detailed}{If TRUE, the dice is plotted with more details}

\item{fill}{Fill color}

\item{fill_success}{Fill color if result is a success}

\item{point_color}{Color of Points}

\item{line_color}{Color of Lines}

\item{width}{Width of dice}

\item{fill}{Fill color}

\item{detailed}{If TRUE, the dice is plotted with more details}
}
\value{
ggplot-Object
}
\description{
Plot result of roll_dice()
}
\examples{
library(magrittr)  
plot_dice()
roll_dice(times = 3, rounds = 3) \%>\% plot_dice()
roll_dice(times = 3, rounds = 3) \%>\% plot_dice(fill_success = "red")
}
