% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice-formula.R
\name{roll_dice_formula}
\alias{roll_dice_formula}
\title{Simulating rolling a dice, using a formula}
\usage{
roll_dice_formula(
  data = NULL,
  dice_formula = "1d6",
  times = 1,
  rounds = 1,
  seed = NULL,
  prob = NULL,
  success = c(6),
  agg = FALSE,
  label = NULL
)
}
\arguments{
\item{data}{Data from a previous experiment}

\item{dice_formula}{}

\item{times}{How many times a dice is rolled (or how many dice are rolled at the same time)}

\item{rounds}{Number of rounds}

\item{seed}{Seed to produce reproducible results}

\item{prob}{Vector of probabilities for each side of the dice}

\item{success}{Which result is a success (default = 6)}

\item{agg}{If TRUE, the result is aggregated (by experiment, rounds) (not implemented)}

\item{label}{Custom text to distinguish an experiment, can be used for plotting etc.}
}
\value{
Result of experiment as a tibble
}
\description{
Example dice_formula:
\itemize{
\item 1d6      > roll one 6-sided dice
\item 1d8      > roll one 8-sided dice
\item 1d12     > roll one 12-sided dice
\item 2d6      > roll two 6-sided dice
\item 1d6e6    > roll one 6-sided dice, explode dice on a 6
\item 3d6kh2   > roll three 6-sided dice, keep top 2 rolls
\item 3d6kl2   > roll three 6-sided dice, keep lowest 2 rolls
\item 4d6kh3e6 > roll four 6-sided dice, keep top 3 rolls, but explode on a 6
\item 1d20+4   > roll one 20-sided dice, and add 4
\item 1d4+1d6  > roll one 4-sided dice and one 6-sided dice, and sum the results
}
}
